+++ core/tauri/src/command.rs
@@ -50,8 +50,9 @@ pub trait CommandArg<'de, R: Runtime>: Sized {
 /// Automatically implement [`CommandArg`] for any type that can be deserialized.
 impl<'de, D: Deserialize<'de>, R: Runtime> CommandArg<'de, R> for D {
   fn from_command(command: CommandItem<'de, R>) -> Result<Self, InvokeError> {
+    let name = command.name;
     let arg = command.key;
-    Self::deserialize(command).map_err(|e| crate::Error::InvalidArgs(arg, e).into())
+    Self::deserialize(command).map_err(|e| crate::Error::InvalidArgs(name, arg, e).into())
   }
 }