+++ core/tauri/src/endpoints/file_system.rs
@@ -9,6 +9,8 @@ use crate::{
 };
 
 use super::InvokeContext;
+#[allow(unused_imports)]
+use anyhow::Context;
 use serde::{
   de::{Deserializer, Error as DeError},
   Deserialize, Serialize,
@@ -127,15 +129,17 @@ impl Cmd {
     context: InvokeContext<R>,
     path: SafePathBuf,
     options: Option<FileOperationOptions>,
-  ) -> crate::Result<Vec<u8>> {
-    file::read_binary(resolve_path(
+  ) -> super::Result<Vec<u8>> {
+    let resolved_path = resolve_path(
       &context.config,
       &context.package_info,
       &context.window,
       path,
       options.and_then(|o| o.dir),
-    )?)
-    .map_err(crate::Error::FailedToExecuteApi)
+    )?;
+    file::read_binary(&resolved_path)
+      .with_context(|| format!("path: {}", resolved_path.0.display()))
+      .map_err(Into::into)
   }
 
   #[module_command_handler(fs_write_file, "fs > writeFile")]
@@ -144,16 +148,18 @@ impl Cmd {
     path: SafePathBuf,
     contents: Vec<u8>,
     options: Option<FileOperationOptions>,
-  ) -> crate::Result<()> {
-    File::create(resolve_path(
+  ) -> super::Result<()> {
+    let resolved_path = resolve_path(
       &context.config,
       &context.package_info,
       &context.window,
       path,
       options.and_then(|o| o.dir),
-    )?)
-    .map_err(Into::into)
-    .and_then(|mut f| f.write_all(&contents).map_err(|err| err.into()))
+    )?;
+    File::create(&resolved_path)
+      .with_context(|| format!("path: {}", resolved_path.0.display()))
+      .map_err(Into::into)
+      .and_then(|mut f| f.write_all(&contents).map_err(|err| err.into()))
   }
 
   #[module_command_handler(fs_read_dir, "fs > readDir")]
@@ -161,23 +167,22 @@ impl Cmd {
     context: InvokeContext<R>,
     path: SafePathBuf,
     options: Option<DirOperationOptions>,
-  ) -> crate::Result<Vec<dir::DiskEntry>> {
+  ) -> super::Result<Vec<dir::DiskEntry>> {
     let (recursive, dir) = if let Some(options_value) = options {
       (options_value.recursive, options_value.dir)
     } else {
       (false, None)
     };
-    dir::read_dir(
-      resolve_path(
-        &context.config,
-        &context.package_info,
-        &context.window,
-        path,
-        dir,
-      )?,
-      recursive,
-    )
-    .map_err(crate::Error::FailedToExecuteApi)
+    let resolved_path = resolve_path(
+      &context.config,
+      &context.package_info,
+      &context.window,
+      path,
+      dir,
+    )?;
+    dir::read_dir(&resolved_path, recursive)
+      .with_context(|| format!("path: {}", resolved_path.0.display()))
+      .map_err(Into::into)
   }
 
   #[module_command_handler(fs_copy_file, "fs > copyFile")]
@@ -186,7 +191,7 @@ impl Cmd {
     source: SafePathBuf,
     destination: SafePathBuf,
     options: Option<FileOperationOptions>,
-  ) -> crate::Result<()> {
+  ) -> super::Result<()> {
     let (src, dest) = match options.and_then(|o| o.dir) {
       Some(dir) => (
         resolve_path(
@@ -194,7 +199,7 @@ impl Cmd {
           &context.package_info,
           &context.window,
           source,
-          Some(dir.clone()),
+          Some(dir),
         )?,
         resolve_path(
           &context.config,
@@ -206,7 +211,8 @@ impl Cmd {
       ),
       None => (source, destination),
     };
-    fs::copy(src, dest)?;
+    fs::copy(src.clone(), dest.clone())
+      .with_context(|| format!("source: {}, dest: {}", src.0.display(), dest.0.display()))?;
     Ok(())
   }
 
@@ -215,7 +221,7 @@ impl Cmd {
     context: InvokeContext<R>,
     path: SafePathBuf,
     options: Option<DirOperationOptions>,
-  ) -> crate::Result<()> {
+  ) -> super::Result<()> {
     let (recursive, dir) = if let Some(options_value) = options {
       (options_value.recursive, options_value.dir)
     } else {
@@ -229,9 +235,11 @@ impl Cmd {
       dir,
     )?;
     if recursive {
-      fs::create_dir_all(resolved_path)?;
+      fs::create_dir_all(&resolved_path)
+        .with_context(|| format!("path: {}", resolved_path.0.display()))?;
     } else {
-      fs::create_dir(resolved_path)?;
+      fs::create_dir(&resolved_path)
+        .with_context(|| format!("path: {} (non recursive)", resolved_path.0.display()))?;
     }
 
     Ok(())
@@ -242,7 +250,7 @@ impl Cmd {
     context: InvokeContext<R>,
     path: SafePathBuf,
     options: Option<DirOperationOptions>,
-  ) -> crate::Result<()> {
+  ) -> super::Result<()> {
     let (recursive, dir) = if let Some(options_value) = options {
       (options_value.recursive, options_value.dir)
     } else {
@@ -256,9 +264,11 @@ impl Cmd {
       dir,
     )?;
     if recursive {
-      fs::remove_dir_all(resolved_path)?;
+      fs::remove_dir_all(&resolved_path)
+        .with_context(|| format!("path: {}", resolved_path.0.display()))?;
     } else {
-      fs::remove_dir(resolved_path)?;
+      fs::remove_dir(&resolved_path)
+        .with_context(|| format!("path: {} (non recursive)", resolved_path.0.display()))?;
     }
 
     Ok(())
@@ -269,7 +279,7 @@ impl Cmd {
     context: InvokeContext<R>,
     path: SafePathBuf,
     options: Option<FileOperationOptions>,
-  ) -> crate::Result<()> {
+  ) -> super::Result<()> {
     let resolved_path = resolve_path(
       &context.config,
       &context.package_info,
@@ -277,7 +287,8 @@ impl Cmd {
       path,
       options.and_then(|o| o.dir),
     )?;
-    fs::remove_file(resolved_path)?;
+    fs::remove_file(&resolved_path)
+      .with_context(|| format!("path: {}", resolved_path.0.display()))?;
     Ok(())
   }
 
@@ -287,7 +298,7 @@ impl Cmd {
     old_path: SafePathBuf,
     new_path: SafePathBuf,
     options: Option<FileOperationOptions>,
-  ) -> crate::Result<()> {
+  ) -> super::Result<()> {
     let (old, new) = match options.and_then(|o| o.dir) {
       Some(dir) => (
         resolve_path(
@@ -295,7 +306,7 @@ impl Cmd {
           &context.package_info,
           &context.window,
           old_path,
-          Some(dir.clone()),
+          Some(dir),
         )?,
         resolve_path(
           &context.config,
@@ -307,7 +318,9 @@ impl Cmd {
       ),
       None => (old_path, new_path),
     };
-    fs::rename(old, new).map_err(crate::Error::Io)
+    fs::rename(&old, &new)
+      .with_context(|| format!("old: {}, new: {}", old.0.display(), new.0.display()))
+      .map_err(Into::into)
   }
 }
 
@@ -318,17 +331,20 @@ fn resolve_path<R: Runtime>(
   window: &Window<R>,
   path: SafePathBuf,
   dir: Option<BaseDirectory>,
-) -> crate::Result<SafePathBuf> {
+) -> super::Result<SafePathBuf> {
   let env = window.state::<Env>().inner();
-  match crate::api::path::resolve_path(config, package_info, env, path, dir) {
+  match crate::api::path::resolve_path(config, package_info, env, &path, dir) {
     Ok(path) => {
       if window.state::<Scopes>().fs.is_allowed(&path) {
         Ok(SafePathBuf(path))
       } else {
-        Err(crate::Error::PathNotAllowed(path))
+        Err(anyhow::anyhow!(
+          crate::Error::PathNotAllowed(path).to_string()
+        ))
       }
     }
-    Err(e) => Err(e.into()),
+    Err(e) => super::Result::<SafePathBuf>::Err(e.into())
+      .with_context(|| format!("path: {}, base dir: {:?}", path.0.display(), dir)),
   }
 }
 
@@ -381,21 +397,21 @@ mod tests {
   #[quickcheck_macros::quickcheck]
   fn read_file(path: SafePathBuf, options: Option<FileOperationOptions>) {
     let res = super::Cmd::read_file(crate::test::mock_invoke_context(), path, options);
-    assert!(!matches!(res, Err(crate::Error::ApiNotAllowlisted(_))));
+    crate::test_utils::assert_not_allowlist_error(res);
   }
 
   #[tauri_macros::module_command_test(fs_write_file, "fs > writeFile")]
   #[quickcheck_macros::quickcheck]
   fn write_file(path: SafePathBuf, contents: Vec<u8>, options: Option<FileOperationOptions>) {
     let res = super::Cmd::write_file(crate::test::mock_invoke_context(), path, contents, options);
-    assert!(!matches!(res, Err(crate::Error::ApiNotAllowlisted(_))));
+    crate::test_utils::assert_not_allowlist_error(res);
   }
 
   #[tauri_macros::module_command_test(fs_read_dir, "fs > readDir")]
   #[quickcheck_macros::quickcheck]
   fn read_dir(path: SafePathBuf, options: Option<DirOperationOptions>) {
     let res = super::Cmd::read_dir(crate::test::mock_invoke_context(), path, options);
-    assert!(!matches!(res, Err(crate::Error::ApiNotAllowlisted(_))));
+    crate::test_utils::assert_not_allowlist_error(res);
   }
 
   #[tauri_macros::module_command_test(fs_copy_file, "fs > copyFile")]
@@ -411,28 +427,28 @@ mod tests {
       destination,
       options,
     );
-    assert!(!matches!(res, Err(crate::Error::ApiNotAllowlisted(_))));
+    crate::test_utils::assert_not_allowlist_error(res);
   }
 
   #[tauri_macros::module_command_test(fs_create_dir, "fs > createDir")]
   #[quickcheck_macros::quickcheck]
   fn create_dir(path: SafePathBuf, options: Option<DirOperationOptions>) {
     let res = super::Cmd::create_dir(crate::test::mock_invoke_context(), path, options);
-    assert!(!matches!(res, Err(crate::Error::ApiNotAllowlisted(_))));
+    crate::test_utils::assert_not_allowlist_error(res);
   }
 
   #[tauri_macros::module_command_test(fs_remove_dir, "fs > removeDir")]
   #[quickcheck_macros::quickcheck]
   fn remove_dir(path: SafePathBuf, options: Option<DirOperationOptions>) {
     let res = super::Cmd::remove_dir(crate::test::mock_invoke_context(), path, options);
-    assert!(!matches!(res, Err(crate::Error::ApiNotAllowlisted(_))));
+    crate::test_utils::assert_not_allowlist_error(res);
   }
 
   #[tauri_macros::module_command_test(fs_remove_file, "fs > removeFile")]
   #[quickcheck_macros::quickcheck]
   fn remove_file(path: SafePathBuf, options: Option<FileOperationOptions>) {
     let res = super::Cmd::remove_file(crate::test::mock_invoke_context(), path, options);
-    assert!(!matches!(res, Err(crate::Error::ApiNotAllowlisted(_))));
+    crate::test_utils::assert_not_allowlist_error(res);
   }
 
   #[tauri_macros::module_command_test(fs_rename_file, "fs > renameFile")]
@@ -448,6 +464,6 @@ mod tests {
       new_path,
       options,
     );
-    assert!(!matches!(res, Err(crate::Error::ApiNotAllowlisted(_))));
+    crate::test_utils::assert_not_allowlist_error(res);
   }
 }