+++ tooling/cli/src/dev.rs
@@ -54,7 +54,7 @@ pub struct Options {
   /// Run the code in release mode
   #[clap(long = "release")]
   release_mode: bool,
-  /// Args passed to the binary
+  /// Command line arguments passed to the runner
   args: Vec<String>,
 }
 
@@ -210,7 +210,10 @@ pub fn command(options: Options) -> Result<()> {
           break;
         }
         if i % 3 == 0 {
-          logger.warn("Waiting for your dev server to start...");
+          logger.warn(format!(
+            "Waiting for your frontend dev server to start on {}...",
+            dev_server_url
+          ));
         }
         i += 1;
         if i == max_attempts {
@@ -225,7 +228,7 @@ pub fn command(options: Options) -> Result<()> {
     }
   }
 
-  let mut process = start_app(&options, &runner, &cargo_features, child_wait_rx.clone());
+  let mut process = start_app(&options, &runner, &cargo_features, child_wait_rx.clone())?;
 
   let (tx, rx) = channel();
 
@@ -268,7 +271,7 @@ pub fn command(options: Options) -> Result<()> {
               break;
             }
           }
-          process = start_app(&options, &runner, &cargo_features, child_wait_rx.clone());
+          process = start_app(&options, &runner, &cargo_features, child_wait_rx.clone())?;
         }
       }
     }
@@ -298,7 +301,7 @@ fn start_app(
   runner: &str,
   features: &[String],
   child_wait_rx: Arc<Mutex<Receiver<()>>>,
-) -> Arc<SharedChild> {
+) -> Result<Arc<SharedChild>> {
   let mut command = Command::new(runner);
   command.args(&["run", "--no-default-features"]);
 
@@ -315,13 +318,13 @@ fn start_app(
   }
 
   if !options.args.is_empty() {
-    command.arg("--").args(&options.args);
+    command.args(&options.args);
   }
 
   command.pipe().unwrap();
 
   let child =
-    SharedChild::spawn(&mut command).unwrap_or_else(|_| panic!("failed to run {}", runner));
+    SharedChild::spawn(&mut command).with_context(|| format!("failed to run {}", runner))?;
   let child_arc = Arc::new(child);
 
   let child_clone = child_arc.clone();
@@ -349,5 +352,5 @@ fn start_app(
     }
   });
 
-  child_arc
+  Ok(child_arc)
 }